+++ src/FsKafka/FsKafka.fs
@@ -91,42 +91,6 @@ type KafkaProducerConfig private (inner, bootstrapServers : string) =
         custom |> Option.iter (fun xs -> for KeyValue (k,v) in xs do c.Set(k,v))
         customize |> Option.iter (fun f -> f c)
         KafkaProducerConfig(c, bootstrapServers)
-    /// Creates and wraps a Confluent.Kafka ProducerConfig with the specified settings
-    [<Obsolete "linger is now mandatory as a result of Confluent.Kafka 1.5's changing the default from 0.5ms to 5ms">]
-    // TODO remove in 2.0.0
-    static member Create
-        (   clientId : string, bootstrapServers : string, acks,
-            /// Message compression. Default: None.
-            ?compression,
-            /// Maximum in-flight requests. Default: 1_000_000.
-            /// NB <> 1 implies potential reordering of writes should a batch fail and then succeed in a subsequent retry
-            ?maxInFlight,
-            /// Time to wait for other items to be produced before sending a batch. Default: 0ms.
-            /// NB the linger setting alone does provide any hard guarantees; see BatchedProducer.Create/ProduceBatch
-            ?linger : TimeSpan,
-            /// Number of retries. Confluent.Kafka default: 2. Default: 60.
-            ?retries,
-            /// Backoff interval. Confluent.Kafka default: 100ms. Default: 1s.
-            ?retryBackoff,
-            /// Statistics Interval. Default: no stats.
-            ?statisticsInterval,
-            /// Ack timeout (assuming Acks != Acks.0). Confluent.Kafka default: 5s.
-            ?requestTimeout,
-            /// Confluent.Kafka default: false. Defaults to true.
-            ?socketKeepAlive,
-            /// Partition algorithm. Default: `ConsistentRandom`.
-            ?partitioner,
-            /// Miscellaneous configuration parameters to be passed to the underlying Confluent.Kafka producer configuration. Same as constructor argument for Confluent.Kafka >=1.2.
-            ?config : IDictionary<string,string>,
-            /// Miscellaneous configuration parameters to be passed to the underlying Confluent.Kafka producer configuration.
-            ?custom,
-            /// Postprocesses the ProducerConfig after the rest of the rules have been applied
-            ?customize) =
-        KafkaProducerConfig.Create(
-            clientId, bootstrapServers, acks, Custom (defaultArg linger (TimeSpan.FromMilliseconds 0.5), defaultArg maxInFlight 1_000_000),
-            ?compression=compression, ?retries=retries, ?retryBackoff=retryBackoff,
-            ?statisticsInterval=statisticsInterval, ?requestTimeout=requestTimeout, ?socketKeepAlive=socketKeepAlive,
-            ?partitioner=partitioner, ?config=config, ?custom=custom, ?customize=customize)
 
 /// Creates and wraps a Confluent.Kafka Producer with the supplied configuration
 type KafkaProducer private (inner : IProducer<string, string>, topic : string) =
@@ -218,8 +182,6 @@ type BatchedProducer private (inner : IProducer<string, string>, topic : string)
         let inner = KafkaProducer.Create(log, config, topic)
         new BatchedProducer(inner.Inner, topic)
 
-#nowarn "44" // TODO remove when obsolete code removed
-
 module Core =
 
     [<NoComparison>]
@@ -238,22 +200,24 @@ module Core =
         member __.InFlightMb = float inFlightBytes / 1024. / 1024.
         member __.Delta(numBytes : int64) = Interlocked.Add(&inFlightBytes, numBytes) |> ignore
         member __.IsOverLimitNow() = Volatile.Read(&inFlightBytes) > maxInFlightBytes
-        [<Obsolete "Please use the overload with the consumer instead">]
-        member __.AwaitThreshold(ct : CancellationToken, busyWork) =
-            if __.IsOverLimitNow() then
-                log.ForContext("maxB", maxInFlightBytes).Information("Consuming... breached in-flight message threshold (now ~{currentB:n0}B), quiescing until it drops to < ~{minMb:n1}MiB",
-                    inFlightBytes, float minInFlightBytes / 1024. / 1024.)
-                while Volatile.Read(&inFlightBytes) > minInFlightBytes && not ct.IsCancellationRequested do
-                    busyWork ()
-                log.Information "Consumer resuming polling"
-        member __.AwaitThreshold(ct : CancellationToken, consumer : IConsumer<_,_>) =
+        member __.AwaitThreshold(ct : CancellationToken, consumer : IConsumer<_,_>, ?busyWork) =
             // Avoid having our assignments revoked due to MAXPOLL (exceeding max.poll.interval.ms between calls to .Consume)
             let showConsumerWeAreStillAlive () =
                 let tps = consumer.Assignment
                 consumer.Pause(tps)
+                match busyWork with Some f -> f () | None -> ()
                 let _ = consumer.Consume(1)
                 consumer.Resume(tps)
-            __.AwaitThreshold(ct, showConsumerWeAreStillAlive)
+            if __.IsOverLimitNow() then
+                log.ForContext("maxB", maxInFlightBytes).Information("Consuming... breached in-flight message threshold (now ~{currentB:n0}B), quiescing until it drops to < ~{minMb:n1}MiB",
+                    inFlightBytes, float minInFlightBytes / 1024. / 1024.)
+                while Volatile.Read(&inFlightBytes) > minInFlightBytes && not ct.IsCancellationRequested do
+                    showConsumerWeAreStillAlive ()
+                log.Information "Consumer resuming polling"
+        [<Obsolete "Please use overload with ?busyWork=None">]
+        // TODO remove ?busyWork=None in internal call when removing this overload
+        member this.AwaitThreshold(ct : CancellationToken, consumer : IConsumer<_,_>) =
+           this.AwaitThreshold(ct, consumer, ?busyWork=None)
 
 /// See https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md for documentation on the implications of specific settings
 [<NoComparison>]
@@ -483,10 +447,10 @@ module private ConsumerImpl =
                         counter.Delta(-batchSize)
                 with e ->
                     log.ForContext("batchSize", batchSize).ForContext("batchLen", batchLen).ForContext("handlerDuration", batchWatch.Elapsed)
-                        .Information(e, "Exiting batch processing loop due to handler exception") 
+                        .Information(e, "Exiting batch processing loop due to handler exception")
                     tcs.TrySetException e |> ignore
                     cts.Cancel() }
-            
+
             let loop = async {
                 use __ = Serilog.Context.LogContext.PushProperty("partition", Binding.partitionValue key.Partition)
                 while not collection.IsCompleted do
@@ -499,7 +463,7 @@ module private ConsumerImpl =
         // run the consumer
         let ct = cts.Token
         try while not ct.IsCancellationRequested do
-                counter.AwaitThreshold(ct, consumer)
+                counter.AwaitThreshold(ct, consumer, ?busyWork=None)
                 try let result = consumer.Consume(ct) // NB TimeSpan overload yields AVEs on 1.0.0-beta2
                     if result <> null then
                         counter.Delta(+approximateMessageBytes result)